resolved Restructure README.md #1
This commit is a first amendment to the original readme, linking to the course's website, acknowledging the license under which the course is originally distributed and brings this repo under the same license as required.
I, unfortunately, have no prior experience with the replit client and am not aware how you wish to use this repo. Should students turn in their course work to you via the pull request? Should they turn it in to Harvard via a pull request?